Replace "wget" with "wget --backups=1"
Without the --backups option wget will store a file as <filename>.1
if the file already exists instead of overwriting it. If you don't
realize this then you think that you downloaded the wrong file.

This is especially important for the *_latest.deb files.

With --backups=1 wget will rename the existing file to <filename>.1
and store the new file as <filename>.